requirements.txt
numpy~=1.25.2
pandas~=2.0.3
matplotlib~=3.7.2
python-dotenv==1.0.0
seaborn
scikit-learn
requests~=2.31.0
geopy
googletrans
xgboost==2.0.2

fastapi
django
sqlalchemy
pydantic
requests
uvicorn
alembic
django-import-export
djangorestframework
